update tests based on TTG-60-change-layout refactor
